Added +sub
+sub (feed name) allows you to subscribe to a feed you've already added with +add. Checks twice a minute. (Requires a subs.csv file, I decided to use a new file so that you didn't have to loop through all the nonsubscribed feeds to check on the subbed ones)
To prevent the subscription showing you posts it's already shown you, it uses the dict lastposts (I didn't want to have to write to a file for every subscription update). Because it's 'in-program' it restarts as the bot does, I actually haven't tested to see what happens if you wait on your subscription like this. Oops. But I know if you wait on your subscription in the same bot-instance as you first subbed to it, it works!

Moved all the code from +get into a function post() and invoked that from +get, so I could invoke post without needing to pass all the context +get needs.

The default feed variable to read from for unknown feeds is now description.

Added delim variable so I can change the delimiter on the fly.
removed some variables I wasn't using and moved some to the commented out section for posterity.

Updated readme and .gitignore.

I haven't necessarily tested the smaller updates, some of them were added in the course of me writing this commit comment, but +sub definitely works :^)