Rewrote feeds storage, added +remove
Feeds are now stored in dicts inside of lists insead of a csv file. (This actually took me a while to figure out- I kept trying to ram a dict into another dict without a key like you can in Lua tables.) Most code in +get, +sub, +add should now be cleaner. Didn't start on a method to actually save the feeds when you close the bot, so you'd have to re+add them after restarting the bot.

+remove now works thanks to the rewrite.

There is no longer a flag in +add to see if posts are expected to be long- The bot should check this automatically now, and some feed descriptions don't normally show a full post, so it will never be long. Depends on the feed.

Updated readme, again. This should probably change with every commit as I refine my roadmap.